requirements.txt
asgiref==3.5.2
Django==4.1.3
django-braces==1.15.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-photologue==3.15.1
django-resized==1.0.2
django-sortedm2m==3.1.1
django-taggit==3.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ExifRead==3.0.0
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
python-slugify==7.0.0
pytz==2022.6
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.3
text-unidecode==1.3
transliterate==1.10.2
tzdata==2022.6
Unidecode==1.3.6
